labs/plaidml_hands_on/requirements.txt
plaidml-keras
